Move some logic from GroundedSAMInvocation to the backend classes.
